Replace xargo with cargo-xbuild
The compiler_builtins crate is now explicitly injected.
When building the alloc crate via a [dependencies.alloc]
key in the Xargo.toml, a “multiple matching crates for compiler_builtins” error occurs.
The reason is that alloc defines its own compiler_builtins dependency.
Thus compiler_builtins in compiled twice, once as version 0.0.0 and once as version 0.1.0.

A workaround is to move to cargo-xbuild.

Signed-off-by: Paolo Teti <paolo.teti@gmail.com>