ESM: add parity fallback handler
When MCU trigger a parity error We can try to recover the VIM table using
a backup table.

Also if an parity error is triggered We need to disable/enable ESM or
FIQ/IRQ

At the moment the VIM table is not restored. I need to dig a bit where
to place the backup table.

Signed-off-by: Paolo Teti <paolo.teti@gmail.com>