ESM: add a way to reset ESM driver
Before jump to main() We need to reset the ESM driver for normal operation
and to load the LTC counter.

If not done just before calling main() this should be done at
the beginning of main(). I've to decide where exactly, so this patch just add
the all the proper functions that are not used yet.

Signed-off-by: Paolo Teti <paolo.teti@gmail.com>