Improve LPO use if offset is too big
If LPO value it too big that would result in a fault due to the
clock test (range check). So, disable range check for a while to be sure
the sudden change will not cause a fault.

Also move some LPO function to flash.rs because We read LPO from OTP memory.